refactor: 엔티티명 변경 FavoriteBoard/JobPost -> InterestBoard/JobPost - 패키지 위치 변경 기존 member 하위 -> interest 패키지 이동
