+++ backend/src/main/java/org/prgrms/devconnect/domain/define/interest/entity/InterestBoard.java
@@ -1,4 +1,4 @@
-package org.prgrms.devconnect.domain.define.member.entity.favoriteboard;
+package org.prgrms.devconnect.domain.define.interest.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,21 +10,22 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.prgrms.devconnect.domain.define.board.entity.Board;
 import org.prgrms.devconnect.domain.define.member.entity.Member;
 
 @Entity
-@Table(name = "favorite_board")
+@Table(name = "interest_board")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FavoriteBoard {
+public class InterestBoard {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "favorite_board_id")
-  private Long favoriteBoardId;
+  @Column(name = "interest_board_id")
+  private Long interestBoardId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "member_id", nullable = false)
@@ -33,4 +34,10 @@ public class FavoriteBoard {
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "board_id", nullable = false)
   private Board board;
+
+  @Builder
+  public InterestBoard(Member member, Board board) {
+    this.member = member;
+    this.board = board;
+  }
 }
