+++ backend/src/main/java/org/prgrms/devconnect/domain/define/interest/entity/InterestJobPost.java
@@ -1,4 +1,4 @@
-package org.prgrms.devconnect.domain.define.member.entity.favoritejobpost;
+package org.prgrms.devconnect.domain.define.interest.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -10,21 +10,22 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.prgrms.devconnect.domain.define.jobpost.entity.JobPost;
 import org.prgrms.devconnect.domain.define.member.entity.Member;
 
 @Entity
-@Table(name = "favorite_job_post")
+@Table(name = "interest_job_post")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FavoriteJobPost {
+public class InterestJobPost {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "favorite_job_post_id")
-  private Long favoriteJobPostId;
+  @Column(name = "interest_job_post_id")
+  private Long interestJobPostId;
 
   @ManyToOne(fetch = FetchType.LAZY)
   @JoinColumn(name = "member_id", nullable = false)
@@ -34,4 +35,9 @@ public class FavoriteJobPost {
   @JoinColumn(name = "job_post_id", nullable = false)
   private JobPost jobPost;
 
+  @Builder
+  public InterestJobPost(Member member, JobPost jobPost) {
+    this.member = member;
+    this.jobPost = jobPost;
+  }
 }
