+++ backend/src/main/java/org/prgrms/devconnect/domain/define/interest/repository/InterestJobPostRepository.java
@@ -1,0 +1,8 @@
+package org.prgrms.devconnect.domain.define.interest.repository;
+
+import org.prgrms.devconnect.domain.define.interest.entity.InterestJobPost;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InterestJobPostRepository extends JpaRepository<InterestJobPost, Long> {
+
+}
